Rework volume implementations
Volumes now always own their resource, and will (later) expose a
view object, which contains a non-owning pointer. They are still
to be though of as shared_ptrs with some extra stuff.

This rework is to avoid sending shared_ptrs to device code, which
is (a) wasteful and (b) may be undefined depending on the
implementation of shared_ptr.