Fix synchronization bug in flat lin. morph. kernel
For line segments which were not axis aligned, threads in a block
could diverge. When that happened, the call to __syncthreads would
block forever. It seems like we don't actually need the explicit
synchronization, so it has been removed, which fixes the bug. If
we end up needing sync. anyway, we can use cooperative groups.