Fix synchronization bug in block processing
What I think would happen: If func would take a long time to exe-
cute, it could happen that the transfer to device blocks would
start before func finished. In this case, the kernels might be
reading from device memory which no longer contained the expected
values. This caused bugs when running on the Linux test machine,
but not on the two Windows machines, for some reason...